ADD: Randomized 10 cubes to appear on app opening
